+++ codex-rs/core/src/lib.rs
@@ -34,6 +34,7 @@ pub mod error;
 pub mod exec;
 pub mod exec_env;
 mod exec_policy;
+pub mod external_agent_config;
 pub mod features;
 mod file_watcher;
 mod flags;
